refactor(music): move the quick action sheets and playlist dialogs into ui/components/music/sheet
